✅  add server test page
