+++ src/components/button.js
@@ -1,6 +1,6 @@
 
 import React from 'react';
-import {StyleSheet, TouchableOpacity} from 'react-native';
+import {StyleSheet, TouchableOpacity, View} from 'react-native';
 
 import {Body} from './typography'
